webapp/requirements.txt
accessify==0.3.1
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
auto-py-to-exe==2.22.0
Babel==2.9.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.24.93
botocore==1.27.93
bottle==0.12.23
bottle-websocket==0.2.9
cached-property==1.5.2
cachetools==5.2.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
colorful==0.5.4
Deprecated==1.2.13
distlib==0.3.6
Django==3.2.15
django-celery-results==2.4.0
django-filter==22.1
djangorestframework==3.14.0
Eel==0.14.0
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.1.2
frozenlist==1.3.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
greenlet==1.1.2
gspread==5.4.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
image==1.5.29
importlib-metadata==4.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.0
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.3.5
pefile==2022.5.30
Pillow==7.1.1
platformdirs==2.5.2
prettyprinter==0.18.0
prompt-toolkit==3.0.32
protobuf==4.21.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive==1.3.1
Pygments==2.12.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pyparsing==3.0.9
PySocks==1.7.1
pyTelegramBotAPI==4.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.6
pywin32-ctypes==0.2.0
PyYAML==6.0
redis==4.3.4
requests==2.28.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
s3transfer==0.6.0
selenium==4.4.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlparse==0.3.1
text-unidecode==1.3
transliterate==1.10.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
virtualenv==20.16.5
wcwidth==0.2.5
Werkzeug==2.1.2
whichcraft==0.6.1
wrapt==1.14.1
wsproto==1.1.0
yarl==1.7.2
zeep==4.1.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
